requirements.txt
django~=3.0.3
psycopg2~=2.8.4
django_imagekit~=4.0.2
django_annoying~=0.10.3
pillow~=7.0.0
django-crispy-forms~=1.9.0
django-taggit~=0.3.1
six~=1.14.0

